// src/app/core/api/generated/schemas/changeUserPasswordDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts API
 * OpenAPI spec version: 1.0
 */

export interface ChangeUserPasswordDto {
  /** Current password. Required when a non-Admin/Manager changes their own password. */
  currentPassword?: string;
  /** @minLength 8 */
  newPassword: string;
}