// src/app/core/api/generated/schemas/paginatedRecentActivityDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts API
 * OpenAPI spec version: 1.0
 */
import type { ActivityItemDto } from "./activityItemDto";
import type { PaginatedRecentActivityDtoMeta } from "./paginatedRecentActivityDtoMeta";

export interface PaginatedRecentActivityDto {
  data: ActivityItemDto[];
  meta: PaginatedRecentActivityDtoMeta;
}