// src/app/core/api/generated/schemas/profileDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { ProfileActivityDto } from "./profileActivityDto";
import type { ProfileAddressDto } from "./profileAddressDto";
import type { ProfileStatsDto } from "./profileStatsDto";
import type { ProfileUserDto } from "./profileUserDto";

export interface ProfileDto {
  user: ProfileUserDto | null;
  address: ProfileAddressDto | null;
  stats: ProfileStatsDto | null;
  activity: ProfileActivityDto;
}