// src/app/core/api/generated/schemas/updateAdminDashboardProfileDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */

export interface UpdateAdminDashboardProfileDto {
  name?: string;
  email?: string;
  phone?: string;
  line1?: string;
  /** @nullable */
  line2?: string | null;
  /** @nullable */
  postalcode?: string | null;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  country?: string | null;
  avatarUrl?: string;
}